changed rank to weight

+++ db/migrate/20160713185904_create_lists.rb
@@ -4,8 +4,8 @@ class CreateLists < ActiveRecord::Migration
       t.string :title
       t.text :description
       t.integer :user_id
-      t.float :pro_sum
-      t.float :con_sum
+      t.integer :pro_sum
+      t.integer :con_sum
       t.timestamps null: false
     end
   end
